qemuarm64-secureboot: set root device correctly
The default root device is /dev/vda, because typically a bare filesystem
is mounted there.

However, qemuarm64-secureboot uses a disk image with partitions, so this
should be /dev/vda2.  Currently this works because qemuarm64-secureboot
has an extlinux.conf which explicitly sets the root device, but we're
working at removing that.

Signed-off-by: Ross Burton <ross.burton@arm.com>
Signed-off-by: Jon Mason <jon.mason@arm.com>