arm/edk2-firmware: add PROVIDES virtual/bootloader
To prevent a collision with u-boot, add the same PROVIDES from it.  The
PROVIDES name need improvement, but this will work in the interim.

This causes a need for making TF-A more flexible.  Add the ability to
reference the UEFI binary for the BL33 portion of the TF-A build
command.  SGI575 is already doing this.  So, it is really just making it
more generic for others to use.

Signed-off-by: Jon Mason <jon.mason@arm.com>