arm-bsp/n1sdp: fix boot due to TF-A 2.7 update
N1SDP board cannot boot after recent TF-A 2.7 update in meta-arm.  This
is due to TF-A 2.7 not configured correctly for N1SDP board to support
trusted boot feature.

This patch temporarily brings back TF-A 2.6 recipes for fixing the N1SDP
boot.

A proper fix is in work progress to configure TF-A 2.7 correctly to
support trutsed boot on N1SDP.

Signed-off-by: Xueliang Zhong <xueliang.zhong@arm.com>
Signed-off-by: Jon Mason <jon.mason@arm.com>