arm-bsp/u-boot: corstone1000: update initramfs bundle size
increase the size of the initramfs bundle used in the boot command

The new trusted services support increases the rootfs size.
When decompressed the initramfs bundle size is around 15M.

u-boot boot command needs to be updated with this size to be able to load
all the initramfs bundle.

When compressed the initramfs bundle size is around 5.4M
(Image.gz-initramfs--5.15.59)

Signed-off-by: Abdellatif El Khlifi <abdellatif.elkhlifi@arm.com>