arm/scp-firmware: rationalise SRCREV/bbappends
In "scp-firmware: allow setting log level to INFO or WARN"[1] the base
SRCREV was bumped from db1991 (the 2.6 tag) to fd7c8356. This means a
fair amount of cleanup can be made:

The base recipe should set PV=2.6+git, both of the existing bbappends
already did this so those can be removed.

The base recipe's CMSIS SRCREV should be a SHA, not a tag name. Both
bbappends also did this so those can be removed.

The TC0 SRCREV is in fact earlier than the new base recipe, so remove
the override so unify the versions. There were no TC-specific changes
between the two SHAs so this should be safe.

[1] 9386db13df5134ab68ae3a1618e999584f7a4261

Signed-off-by: Ross Burton <ross.burton@arm.com>
Signed-off-by: Jon Mason <jon.mason@arm.com>