meta-arm: add build to gitignore
To avoid having always tools that depend on git ls or other
git plumbing to include and spin around the enormous content of
the build directory.

Just add it to the ignore file and make that build content,
that will never get in the repo invisible to git and tools.

Signed-off-by: Rui Miguel Silva <rui.silva@linaro.org>
Signed-off-by: Jon Mason <jon.mason@arm.com>