arm/trusted-services: use apply_local_src_patches
Now that the apply_local_src_patches class supports being used with
multiple directories, use that instead of reimplementing the logic.

Also remove redundant patchdir assignments as these patches are against
the trusted-services repository, which is ${S}.  I suspect these are
exposing a subtle bug in the core patching logic which meant the local
patches were not applying correctly.

Signed-off-by: Ross Burton <ross.burton@arm.com>
Signed-off-by: Jon Mason <jon.mason@arm.com>